Added 'Age' parameter to control spawning, aging, other fiixes

+++ src/main/java/com/animania/common/entities/pigs/ai/EntityAIPigSnuffle.java
@@ -2,6 +2,7 @@ package com.animania.common.entities.pigs.ai;
 
 import com.animania.common.handler.ItemHandler;
 import com.animania.common.helper.ItemHelper;
+import com.animania.config.AnimaniaConfig;
 import com.google.common.base.Predicate;
 import com.google.common.base.Predicates;
 
@@ -74,7 +75,10 @@ public class EntityAIPigSnuffle extends EntityAIBase
 
 			if (EntityAIPigSnuffle.IS_TALL_GRASS.apply(this.entityWorld.getBlockState(blockpos))) {
 
-				this.entityWorld.destroyBlock(blockpos, false);
+				if (this.entityWorld.getGameRules().getBoolean("mobGriefing") && AnimaniaConfig.gameRules.plantsRemovedAfterEating) {
+					this.entityWorld.destroyBlock(blockpos, false);
+				}
+				
 				this.grassEaterEntity.eatGrassBonus();
 			}
 			else {
